refactor-#35: move repository and entity to infrastructure
