+++ src/catalog/infrastructure/entities/activity.entity.ts
