+++ src/catalog/infrastructure/infrastructure.module.ts
@@ -1,4 +1,11 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-@Module({})
+import { ActivityEntity } from './entities';
+import { ActivityRepository } from './repositories';
+
+@Global()
+@Module({
+    imports:[TypeOrmModule.forFeature([ActivityRepository, ActivityEntity])],
+})
 export class InfrastructureModule {}
