+++ src/catalog/infrastructure/repositories/activity.repository.ts
